Refactor with BigDecimal and ResponseEntity

+++ src/main/java/com/leanpay/loancalculator/model/response/amortization/AmortizationResponse.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Document
@@ -13,6 +14,6 @@ import java.util.List;
 public class AmortizationResponse {
     @Id
     private String id;
-    private double loanAmount;
+    private BigDecimal loanAmount;
     private List<AmortizationSchedule> amortizationSchedule;
 }
