+++ src/main/java/com/leanpay/loancalculator/model/response/amortization/AmortizationSchedule.java
@@ -3,12 +3,14 @@ package com.leanpay.loancalculator.model.response.amortization;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class AmortizationSchedule {
     private int id;
-    private double paymentAmount;
-    private double principalAmount;
-    private double interestAmount;
-    private double balanceOwed;
+    private BigDecimal paymentAmount;
+    private BigDecimal principalAmount;
+    private BigDecimal interestAmount;
+    private BigDecimal balanceOwed;
 }
